Documentatie comments toegevoegd.

+++ src/main/java/com/marktplaats/controller/GebruikerController.java
@@ -33,6 +33,7 @@ public class GebruikerController {
         this.productService = productService;
     }
 
+    /**Methode om de gebruiker te updaten.*/
     @PutMapping(value = "/update")
     public ResponseEntity<Object> UpdateGebruiker(@RequestBody Gebruiker gebruiker,@AuthenticationPrincipal OAuth2User user){
         String id = user.getAttributes().get("sub").toString();
@@ -40,7 +41,8 @@ public class GebruikerController {
         if(this.gebruikerService.FindById(id) != null && AllFieldsAreFilled(gebruiker)){
             gebruiker.setEmail(email);
             gebruiker.setId(id);
-            if(this.gebruikerService.GebruikersnaamAlInGebruik(gebruiker.getGebruikersnaam())){
+            boolean same = gebruiker.getId() == this.gebruikerService.FindByGebruikersnaam(gebruiker.getGebruikersnaam()).getId();
+            if(this.gebruikerService.GebruikersnaamAlInGebruik(gebruiker.getGebruikersnaam()) && !same){
                 return new ResponseEntity<>(gson.toJson("Gebruikersnaam is al in gebruik."),HttpStatus.BAD_REQUEST);
             }
             else{
@@ -51,11 +53,13 @@ public class GebruikerController {
         return new ResponseEntity<>(gson.toJson("Er ging iets mis."),HttpStatus.BAD_REQUEST);
     }
 
+    /**controleert of alle attributen van Gebruiker zijn ingevoerd.*/
     public boolean AllFieldsAreFilled(Gebruiker gebruiker){
         return gebruiker.getAchternaam() != null  && gebruiker.getGebruikersnaam() != null &&
                 gebruiker.getGeboorteDatum() != null && gebruiker.getGeslacht() != null && gebruiker.getVoornaam() != null;
     }
 
+    /**Methode die alle producten ophaalt van de meegegeven gebruiker.*/
     @GetMapping(value = "/producten")
     public ResponseEntity<Object> GetAllProducten(@AuthenticationPrincipal OAuth2User user){
         String id = user.getAttributes().get("sub").toString();
@@ -67,12 +71,14 @@ public class GebruikerController {
         return new ResponseEntity<>(products, HttpStatus.OK);
     }
 
+    /**Methode die alle gebruikers ophaalt.*/
     @GetMapping(value = "/getAll")
     public ResponseEntity<Object> GetAll(){
         List<Gebruiker> gebruikers = gebruikerService.GetAll();
         return new ResponseEntity<>(gebruikers,HttpStatus.OK);
     }
 
+    /**Methode die een product aanmaakt en koppelt aan de gebruiker die de product aanmaakt.*/
     @PostMapping(value = "/voegProductToe")
     public ResponseEntity<Object> VoegProductToe(@RequestBody GebruikerProducten gebruikerProducten){
         if(Objects.nonNull(gebruikerProducten.gebruiker) || Objects.nonNull(gebruikerProducten.product)){
@@ -83,6 +89,7 @@ public class GebruikerController {
                 ,HttpStatus.BAD_REQUEST);
     }
 
+    /**Methode die een product verwijderd van de meegegeven gebruiker.*/
     @DeleteMapping(value = "/deleteProduct/{id}")
     public ResponseEntity<Object> VerwijderProduct(@PathVariable int id, @AuthenticationPrincipal OAuth2User user){
         String userId = user.getAttributes().get("sub").toString();
@@ -97,17 +104,21 @@ public class GebruikerController {
                 ,HttpStatus.BAD_REQUEST);
     }
 
+    /**Methode die de meegegeven gebruiker verwijdert,
+     * nadat er is gecontroleerd of de te verwijderen gebruiker wel dezelfde gebruiker is als vanwaar de request vandaan komt*/
     @DeleteMapping(value = "/delete/{id}")
     public ResponseEntity<Object> VerwijderGebruiker(@PathVariable String id,@AuthenticationPrincipal OAuth2User user){
         String userId = user.getAttributes().get("sub").toString();
-        if(this.gebruikerService.FindById(id) != null && this.gebruikerService.FindById(userId) != null){
-            List<GebruikerProducten> producten = this.gebruikerService.GetAllProducten(id);
-            this.gebruikerService.VerwijderAlleProducten(id);
-            for (var product:producten) {
-                this.productService.Delete(product.getProduct());
+        if(id == userId){
+            if(this.gebruikerService.FindById(id) != null && this.gebruikerService.FindById(userId) != null){
+                List<GebruikerProducten> producten = this.gebruikerService.GetAllProducten(id);
+                this.gebruikerService.VerwijderAlleProducten(id);
+                for (var product:producten) {
+                    this.productService.Delete(product.getProduct());
+                }
+                this.gebruikerService.Delete(this.gebruikerService.FindById(id));
+                return new ResponseEntity<>(gson.toJson("Gebruiker is succesvol verwijderd."),HttpStatus.OK);
             }
-            this.gebruikerService.Delete(this.gebruikerService.FindById(id));
-            return new ResponseEntity<>(gson.toJson("Gebruiker is succesvol verwijderd."),HttpStatus.OK);
         }
         return new ResponseEntity<>(gson.toJson("Er is iets misgegaan."),HttpStatus.BAD_REQUEST);
     }
